refaktoryzacja
