+++ frontend/src/components/AdminRoute.jsx
@@ -1,23 +1,12 @@
 import { Navigate } from 'react-router-dom';
 import { useAuth } from '../context/AuthContext';
-import { Spin } from 'antd';
+import LoadingSpinner from './LoadingSpinner';
 
 const AdminRoute = ({ children }) => {
   const { user, isAuthenticated, loading } = useAuth();
 
   if (loading) {
-    return (
-      <div
-        style={{
-          display: 'flex',
-          justifyContent: 'center',
-          alignItems: 'center',
-          minHeight: '60vh',
-        }}
-      >
-        <Spin size='large' />
-      </div>
-    );
+    return <LoadingSpinner />;
   }
 
   if (!isAuthenticated || user?.role !== 'admin') {
