+++ frontend/src/components/FormField.module.css
@@ -1,0 +1,34 @@
+.inputGroup {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 24px;
+}
+
+.inputGroup label {
+  color: var(--color-blue);
+  font-size: 16px;
+  margin-bottom: 8px;
+  font-weight: 500;
+}
+
+.inputGroup input {
+  background-color: var(--color-grey-bg);
+  border: 2px solid var(--color-blue);
+  border-radius: 8px;
+  padding: 12px 16px;
+  font-family: var(--font-family);
+  font-size: 16px;
+  outline: none;
+  color: var(--color-black);
+}
+
+.inputGroup input:focus {
+  box-shadow: 0 0 0 2px rgba(71, 166, 255, 0.2);
+}
+
+.errorMessage {
+  color: #ff4d4f;
+  font-size: 14px;
+  margin-top: 4px;
+  font-weight: 700;
+}
