+++ frontend/src/layouts/Layout.module.css
@@ -1,0 +1,10 @@
+.layoutWrapper {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+
+.contentArea {
+  flex: 1;
+  padding-top: 68px;
+}
