+++ frontend/src/layouts/NavbarOnlyLayout.jsx
@@ -1,13 +1,12 @@
 import { Outlet } from 'react-router-dom';
 import Navbar from '../components/Navbar.jsx';
+import styles from './Layout.module.css';
 
 const NavbarOnlyLayout = () => {
   return (
-    <div
-      style={{ display: 'flex', flexDirection: 'column', minHeight: '100vh' }}
-    >
+    <div className={styles.layoutWrapper}>
       <Navbar />
-      <div style={{ flex: 1, paddingTop: '68px' }}>
+      <div className={styles.contentArea}>
         <Outlet />
       </div>
     </div>
