+++ frontend/src/pages/Home.module.css
@@ -3,7 +3,7 @@
   margin-right: 80px;
   margin-top: 40px;
   margin-bottom: 40px;
-  background-color: linear;
+
 }
 
 .titleStart {
@@ -87,11 +87,7 @@
   rotate: -135deg;
 }
 
-.offersIconRight {
-  color: var(--color-blue);
-  font-size: 40px;
-  rotate: 45deg;
-}
+
 
 .howItWorksSection {
   text-align: center;
