+++ frontend/src/pages/Register.jsx
@@ -2,8 +2,14 @@ import styles from './Register.module.css';
 import { useForm } from 'react-hook-form';
 import { Typography, Row, Col } from 'antd';
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import Button from '../components/Buttons';
+import FormField from '../components/FormField';
+import apiClient from '../api/apiClient';
+import {
+  PASSWORD_REGEX,
+  PASSWORD_REQUIREMENTS_MESSAGE,
+} from '../utils/constants';
 
 const { Title, Paragraph } = Typography;
 
@@ -20,24 +26,9 @@ const Register = () => {
     reValidateMode: 'onBlur',
   });
 
-  const handleRegister = async (data) => {
+  const handleRegister = async (formValues) => {
     try {
-      const response = await fetch('http://localhost:3000/api/auth/register', {
-        method: 'POST',
-        body: JSON.stringify(data),
-        headers: {
-          'Content-Type': 'application/json',
-          Accept: 'application/json',
-        },
-        credentials: 'include',
-      });
-      const responseData = await response.json();
-
-      if (!response.ok) {
-        throw new Error(
-          responseData.error || 'Wystąpił błąd podczas rejestracji'
-        );
-      }
+      await apiClient.post('/auth/register', formValues);
       setSuccess(true);
       setTimeout(() => {
         navigate('/login');
@@ -56,7 +47,8 @@ const Register = () => {
         <Col xs={24} md={12} className={styles.leftSide}>
           <div className={styles.overlayBox}>
             <Title level={2} className={styles.overlayTitle}>
-              Technologie aplikacji <br /> webowych
+              Technologie aplikacji <br />
+              webowych
             </Title>
             <Title level={3} className={styles.overlaySubtitle}>
               Rezerwacja obiektów
@@ -78,86 +70,56 @@ const Register = () => {
             </Title>
 
             <form noValidate onSubmit={handleSubmit(handleRegister)}>
-              <div className={styles.inputGroup}>
-                <label htmlFor='firstName'>Imię</label>
-                <input
-                  type='text'
-                  id='firstName'
-                  {...register('firstName', {
-                    required: 'Musisz podać imię',
-                    minLength: {
-                      value: 2,
-                      message: 'Imie musi miec co najmniej 2 znaki',
-                    },
-                  })}
-                />
-                {errors.firstName && (
-                  <div className={styles.errorMessage}>
-                    {errors.firstName.message}
-                  </div>
-                )}
-              </div>
-
-              <div className={styles.inputGroup}>
-                <label htmlFor='lastName'>Nazwisko</label>
-                <input
-                  type='text'
-                  id='lastName'
-                  {...register('lastName', {
-                    required: 'Musisz podać nazwisko',
-                    minLength: {
-                      value: 2,
-                      message: 'Nazwisko musi miec co najmniej 2 znaki',
-                    },
-                  })}
-                />
-                {errors.lastName && (
-                  <div className={styles.errorMessage}>
-                    {errors.lastName.message}
-                  </div>
-                )}
-              </div>
-
-              <div className={styles.inputGroup}>
-                <label htmlFor='email'>Email</label>
-                <input
-                  type='email'
-                  id='email'
-                  {...register('email', {
-                    required: 'Musisz podać email',
-                    pattern: {
-                      value: /[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}/,
-                      message: 'Nieprawidłowy format emaila',
-                    },
-                  })}
-                />
-                {errors.email && (
-                  <div className={styles.errorMessage}>
-                    {errors.email.message}
-                  </div>
-                )}
-              </div>
-
-              <div className={styles.inputGroup}>
-                <label htmlFor='password'>Hasło</label>
-                <input
-                  type='password'
-                  id='password'
-                  {...register('password', {
-                    required: 'Musisz podać hasło',
-                    pattern: {
-                      value: /^(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z0-9]).{8,}$/,
-                      message:
-                        'Nieprawidłowy format hasła (wymagane 8 znaków, wielka litera, cyfra i znak specjalny)',
-                    },
-                  })}
-                />
-                {errors.password && (
-                  <div className={styles.errorMessage}>
-                    {errors.password.message}
-                  </div>
-                )}
-              </div>
+              <FormField
+                label='Imię'
+                id='firstName'
+                error={errors.firstName?.message}
+                register={register('firstName', {
+                  required: 'Musisz podać imię',
+                  minLength: {
+                    value: 2,
+                    message: 'Imię musi mieć co najmniej 2 znaki',
+                  },
+                })}
+              />
+              <FormField
+                label='Nazwisko'
+                id='lastName'
+                error={errors.lastName?.message}
+                register={register('lastName', {
+                  required: 'Musisz podać nazwisko',
+                  minLength: {
+                    value: 2,
+                    message: 'Nazwisko musi mieć co najmniej 2 znaki',
+                  },
+                })}
+              />
+              <FormField
+                label='Email'
+                id='email'
+                type='email'
+                error={errors.email?.message}
+                register={register('email', {
+                  required: 'Musisz podać email',
+                  pattern: {
+                    value: /[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}/,
+                    message: 'Nieprawidłowy format emaila',
+                  },
+                })}
+              />
+              <FormField
+                label='Hasło'
+                id='password'
+                type='password'
+                error={errors.password?.message}
+                register={register('password', {
+                  required: 'Musisz podać hasło',
+                  pattern: {
+                    value: PASSWORD_REGEX,
+                    message: PASSWORD_REQUIREMENTS_MESSAGE,
+                  },
+                })}
+              />
 
               {errors.root && (
                 <div className={styles.errorMessage}>{errors.root.message}</div>
@@ -179,7 +141,7 @@ const Register = () => {
             </form>
 
             <div className={styles.loginLink}>
-              Posiadasz już konto? <a href='/login'>Zaloguj się</a>
+              Posiadasz już konto? <Link to='/login'>Zaloguj się</Link>
             </div>
           </div>
         </Col>
